requirements.txt

CherryPy==18.1.1
psutil==5.6.1
mdstat[linuxraid]
mdstat==1.0.4
netaddr==0.7.19
itsdangerous==1.1.0
python-pam==1.8.4
Mako==1.0.9
uptime==3.0.1
py-cpuinfo==5.0.0
